+++ advancedDungeon/src/produsAdvanced/abstraction/FireballSkill.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.math.MathUtils;
 import contrib.components.ItemComponent;
-import contrib.skill.DamageProjectileSkill;
 import contrib.skill.Skill;
+import contrib.skill.projectileSkill.DamageProjectileSkill;
 import contrib.utils.components.health.DamageType;
 import core.Entity;
 import core.utils.Point;
