+++ devDungeon/src/entities/BossAttackSkills.java
@@ -4,10 +4,10 @@ import contrib.components.CollideComponent;
 import contrib.components.HealthComponent;
 import contrib.components.SpikyComponent;
 import contrib.entities.AIFactory;
-import contrib.skill.DamageProjectileSkill;
-import contrib.skill.FireballSkill;
 import contrib.skill.Skill;
 import contrib.skill.SkillTools;
+import contrib.skill.projectileSkill.DamageProjectileSkill;
+import contrib.skill.projectileSkill.FireballSkill;
 import contrib.systems.EventScheduler;
 import contrib.utils.EntityUtils;
 import contrib.utils.components.health.DamageType;
