+++ devDungeon/src/entities/MonsterType.java
@@ -9,9 +9,9 @@ import contrib.entities.MonsterDeathSound;
 import contrib.entities.MonsterFactory;
 import contrib.entities.MonsterIdleSound;
 import contrib.hud.DialogUtils;
-import contrib.skill.FireballSkill;
 import contrib.skill.SkillTools;
-import contrib.skill.TPBallSkill;
+import contrib.skill.projectileSkill.FireballSkill;
+import contrib.skill.projectileSkill.TPBallSkill;
 import contrib.utils.components.ai.fight.AIChaseBehaviour;
 import contrib.utils.components.ai.fight.AIRangeBehaviour;
 import contrib.utils.components.ai.idle.PatrolWalk;
