+++ devDungeon/src/level/utils/LevelUtils.java
@@ -1,6 +1,6 @@
 package level.utils;
 
-import contrib.skill.TPBallSkill;
+import contrib.skill.projectileSkill.TPBallSkill;
 import contrib.systems.EventScheduler;
 import core.Game;
 import core.level.Tile;
