+++ dungeon/src/contrib/item/concreteItem/ItemWoodenBow.java
@@ -4,8 +4,8 @@ import contrib.components.InventoryComponent;
 import contrib.configuration.KeyboardConfig;
 import contrib.entities.WorldItemBuilder;
 import contrib.item.Item;
-import contrib.skill.BowSkill;
 import contrib.skill.SkillTools;
+import contrib.skill.projectileSkill.BowSkill;
 import core.Entity;
 import core.Game;
 import core.components.InputComponent;
