+++ dungeon/src/contrib/skill/projectileSkill/TPBallSkill.java
@@ -1,4 +1,4 @@
-package contrib.skill;
+package contrib.skill.projectileSkill;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.audio.Sound;
